todo-api/requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
cffi==1.14.0
click==7.1.1
cryptography==2.9
Flask==1.1.2
Flask-HTTPAuth==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pathspec==0.8.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
regex==2020.4.4
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.1
